Always write a done file
The master node will now write a done file when it has finished writing the results.
The contents of the done file will be the name of the results file that sonobuoy has written.

The idea here is that you can now react to the results being generated using a second
container that knows to look for a `done` file in some directory.

An example of this would be uploading the results to s3 after sonobuoy master finishes.

Signed-off-by: Chuck Ha <chuck@heptio.com>