Implement the done file interface
Once Sonobuoy is finished running it writes the path of the results tarball
into a done file located in the results directory (as configured via
the config file or an env var).

The goal of this is to be able to react to a finished sonobuoy run and
do something interesting with the results. For instance, one could
upload the results to S3 for permanent storage and need not create a long
lived volume for sonobuoy specifically.

Signed-off-by: Chuck Ha <chuck@heptio.com>