[WIP] Change how plugins work
Instead of creating our own schema for a plugin we will use annotations
to get metadata to sonobuoy instead. We can then define full kubernetes
resources as plugins.

The resources we define are templates. The idea is to eventually move
these templates over to ksonnet and have a nice ksonnet library for the
shared pieces between plugins.

Remaining work:

* Whatever configuration can be passed into the template
should be allowed to be passed into the template (`RESULTS_DIR`).
* Add a test or two to the plugin.Interface implementations

Signed-off-by: Chuck Ha <chuck@heptio.com>