Acknowledge state between plugins complete and done
If your plugin artifacts are large then it will take
a few seconds for the tarball to be created. However,
the aggregator marks the pod status as done and
therefore there can be a few seconds where the
status reports that results are ready for download
but you will wither get a "file does not exist" error
or a tarball locally that is only partially complete
and most likely corrupted.

Fixes: #572
Signed-off-by: John Schnake <jschnake@vmware.com>