List files created by retrieve command
When the user calls sonobuoy retrieve they are given no output
and then left to figure out if they got any files and what their
name is.

As someone who runs this a lot I'd like to be able to know more
easily what files were created so that I can script around them
or more easily grep for the "new" file in a sea of older tarballs.

Fixes #594

Signed-off-by: John Schnake <jschnake@vmware.com>