Update README
This changes the README in a few main ways:
 - Adds more details about quick mode which is often
used not but front and center in the README
 - Adds improved script lines which take advantate
of --wait options and capturing the filename from
retrieve
 - Adds information regarding `sonobuoy gen` which is the
best workaround whenever Sonobuoy lacks functionality users
need. Slack is usually full of questions where the answer is:
use `sonobuoy gen` and `kubectl apply`
 - Moves the link definitions to a central location at the
bottom of the file.
 - Removes prefix from code lines to improve copy/paste use
 - Clarifies installation can also be from binary which
doesnt require Go and that kubectl isn't required except in
some workflows

Fixes #614

Signed-off-by: John Schnake <jschnake@vmware.com>