Default to 'auto' for gen command
There exists a bifurcation in the kube-conformance logic
which is unintuitive and potentially damaging to the user
experience.

Currently, if they use sonobuoy gen the kube-conformance
version is defaulted to `latest` but if they use sonobuoy run
it is defaulted to `auto`. The problem is that one may work
on a cluster while the other will not: leaving users in a
sutiation where run will work but gen -> kubectl apply will
not without an indication why.

This change causes the two to be in sync. The downside is that
there are cases where you may run sonobuoy gen without a
current KUBECONFIG, so now sonobuoy gen will fail.

The error is descriptive and the workaround is to provide the
image version you want explicitly rather than falling back to
latest implicitly.

I think this is preferred behavior, since some users would have
already been broken with the existing logic.

Fixes: #619
Signed-off-by: John Schnake <jschnake@vmware.com>