Replace alpine with debian-slim
Updates the images we use when building Sonobuoy.
For the build image this also updates go from
1.12.0 to 1.12.1.

Due to the fact that golang doesnt offer a
stretch-slim image, the 1.12.1 stretch image
is used.

Stretch slim seems to come with what it needs
so the dockerfile CMD1 can just be removed. For
ease of future changes I left the CMD1 in place and,
just like on arm64, we just remove the line rather
than replace it.

One caveat to this update was that differences in
how archive/tar and/or the platform tar implementation
works w.r.t. the end-of-archive marker caused the
debian images to make `sonobuoy retrieve` hand. A
workaround had to be added to ensure any trailing
0s were read.

Fixes #609

Signed-off-by: John Schnake <jschnake@vmware.com>