Use workers default image and pull IfNotPresent
The worker's default image was using a hardcoded
value and not referencing the const that was
defined for that purpose leading to the
aggregator and worker using different images.

In addition, changing the default pull policy
to IfNotPresent seems like it could only help
in certain situations and is low risk otherwise.

Fixes: #643
Signed-off-by: John Schnake <jschnake@vmware.com>