Fix CRD query
The CRD query did not have a properly formed
rest client leading to the wrong URLs being
used to query for CRDs. This resulted in an
error every Sonobuoy run on the aggregator.

This change passes the rest.Config itself
to the method so that we can create any
clients we need from it. That will be useful
later if we want to move to the dynamic client
for all queries.

Fixes #640

Signed-off-by: John Schnake <jschnake@vmware.com>