Add validation to focus/skip regexp flags
When creating a run and using the focus/skip
flags it can be common to accidently provide
an invalid regexp such as improperly escaping
certain characters or using `*` instead of
`.*` This change ensures that we are validing
and failing early to create a better user
experience.

Currently, if the user does this the sonobuoy
run will get created and the e2e plugin will
fail but you have to retrieve the output and
inspect the logs manually to see the panic
from the framework.

Fixes #615

Signed-off-by: John Schnake <jschnake@vmware.com>