Support imagePullSecrets for sonobuoy workers
In the situation where the sonobuoy image itself was
in a private registry there was no recourse for users
because imagePullSecrets was not defined in the
config object used to help populate the templates
and these templates were generated on the server
side (so the user couldn't use sonobuoy gen and
manually manipulate them).

This change does the following:
 - adds imagePullSecrets to the config type
 - adds sections in the templates to honor it
including all the plugins and the aggregator.

This at least unblocks people who may be in
this situation by allowing them to specify the value
in their custom sonobuoy config.

Fixes #424

Signed-off-by: John Schnake <jschnake@vmware.com>