Add plugin flag for custom, local plugins
The user can now load custom plugin manifests into the
YAML that Sonobuoy will run with the --plugin flag.

For backwards compatibility, if no plugins are specified
in this manner and instead a PluginSelection object is
still specified, we will fallback to adding the necessary
plugins.

When loading plugins using the flag, they should specify
either a filename which points to a plugin manifest or
one of the sentinal values representing our internal plugins:
e2e or systemd-logs.

Plugins must be uniquely defined by the plugin name.

Fixes #405

Signed-off-by: John Schnake <jschnake@vmware.com>