Change logic in template to use a list of plugin manifests
Currently, the template for the initial YAML (via sonobuoy gen)
has it hardcoded that both the e2e and systemd-logs plugin are
configured. This is 1) misleading to users who might see this and
think both are going to be run and 2) brittle.

With upcoming changes wanting to make the plugin list more
variable, it makes sense to put this change first: go ahead and
track a plugin list and adjust the template to iterate over the
plugins, inserting their YAML each time.

This makes the YAML more clear, flexible, and in some cases shorter.

Fixes #672

Signed-off-by: John Schnake <jschnake@vmware.com>