Defer assignment of Docker gcr.io secret
With the ':=' assignment in the Makefile, the SECRET was
evaluated at the start, regardless of whether it was needed
or not to complete the recipe.

This causes an issue if you don't have the necessary
docker-credential-gcr tool which is expected to be there
for Jenkins.

This change defers the evaluation by using '=' instead and
avoids confusing error messaages when building locallly.

Signed-off-by: John Schnake <jschnake@vmware.com>