Tighten preflight check to err on existing namespace
The typical flow for Sonobuoy involves creating a new namespace
and it is uncommon to need to put it into an existing namespace.
When that occurs, cleanup must be manual since Sonobuoy by default
just deletes the namespace.

This tightens the precheck to error on the existing namespace. In
situations where that is still desired, preflight checks can be
skipped.

Refactored the logic some so that tests could be added for preflight
checks.

Fixes #677

Signed-off-by: John Schnake <jschnake@vmware.com>