Fix flaky unit test
The ordering of YAML lists is not well defined and so this passed
tests locally and in CI by chance. Even trying to repeat the
flakiness failed at first because I was hitting the testing cache.

Removing the 2nd end var does little to limit the value of the test
and we also have functional tests (e2e integation tests) which
will hit this flow.

Signed-off-by: John Schnake <jschnake@vmware.com>