Add sonobuoy gen config file and documentation
The Sonobuoy config file is a little mysterious to most
users since the values are only really defined in code
and not well documented.

This change adds a command to output the default config
so that users might tweak the values they need more
easily.

In addition, I updated the documentation to make it more
clear for everyone what the configuration file is and
how each of the files are used.

Fixes #661

Signed-off-by: John Schnake <jschnake@vmware.com>