Add test that will run through command creation
The CLI code has very low code coverage which indicates
that it has lots of areas potentially containing bugs.

However, a large chunk of the code in the CLI is just
boilerplate that sets up cobra commands and doesn't
execute much actual logic.

The point of this test is to have a code flow go through
those so that code coverage better reflects actual,
substantive code.

Relates to #651

Signed-off-by: John Schnake <jschnake@vmware.com>