Run all loaded plugins by default
The aggregator has a plugin filtering mechanism that
seems to actually complicate things for most users. Beyond
adding their plugin to the config map, they also have to
adjust the plugin selection.

This seems to have originated due to a desire for
flexibility (before knowing we needed it or not) and also
ease of testing.

This change just introduces special handling for the nil
selection (not empty, completely nil). In that case, all
plugins are run which are loaded.

ref #405

Signed-off-by: John Schnake <jschnake@vmware.com>