Modify the json serialization of config object
Makes the json serialization omit the LoadedPlugins
field which is something the object needs when running
but is not part of the config itself.

Also makes the plugin selection list omitempty since
we would like to have the null value indicate "run all"
in which case most users do not need to see or edit
that value.

ref #405

Signed-off-by: John Schnake <jschnake@vmware.com>