Update queryies to use the dynamic client
This change is fairly substantial so I wanted to specify
a few details:
 - All queries are now executed with the discovery/dynamic
clients walking the API server
 - QPS and Burst can be set on the config.json so that you can
speed up/slow down your queries based on speed and throttling.
 - The filenames containing the the resource query results are
updated to list group_version_resource.json as opposed to
resource.json. This makes it explicitly clear where the resources
came from and avoids a problem where different groups with the
same resource name could collide.
 - To avoid collision in general though, the queries only grab
the first resource with a given name that it encounters. This
is the behavior kubectl uses as well.
 - Resource filtering via the config.json field Resource
is still done, but explicitly compared against the resource.Name
field given by the server. Most visibly, the casing of the values
is different than it used to be.
 - If an nil Resource value is set in the config.json, all values
will be queried with the exception of secrets.
 - If an empty array of Resources is set in the config.json, no
values will be queried, same as today.
 - The default resource list exists but has a few extra values
added to it. In the future we may remove it entirely and just
default to query all. This was kept so make the casing changes
more clear to users of sonobuoy gen.

Fixes #492
Fixes #630

Signed-off-by: John Schnake <jschnake@vmware.com>