Add example for command runner plugin
With the new plugin generation/runner capabilities we needed
a new example to show users what they can do/how to do it.

This example just takes all the arguments it is given and
runs them with /bin/sh, saves the output, tars it up, and
tells the Sonobuoy worker where to find the results.

This example includes:
 - the code for the plugin (a single bash script)
 - the Dockerfile to build the script
 - a README.md to explain what it does and how to use it

Signed-off-by: John Schnake <jschnake@vmware.com>