Add documentation about how to pull private Sonobuoy image
The process to use a private Sonobuoy image is non-trivial since
you have to create a secret in the same namespace.

In the future, if this is common, we can simplify the process but
currently this documentation should clarify it for the few people
who need to use this flow.

Fixes #692

Signed-off-by: John Schnake <jschnake@vmware.com>