Set any plugin env vars from the CLI
Added a new flag and backing type in order to gather and
set env vars on any plugin from the command line which
makes it easier to iterate on plugins/tests.

This PR does the following:
 - adds new flag --plugin-env and backing type
 - adds new field on GenConfig in order to pass this to
pkg/client/gen/GenerateManifest(...) in order to apply
the changes after the dynamic plugins have been generated.
 - adds logic to sort the env var to ensure a stable order
for testing/use

On the command line the values are of the form:
--plugin-env pluginName.env=val

Fixes #639

Signed-off-by: John Schnake <jschnake@vmware.com>