Ensure conformance image version is prefixed with 'v'
When using flags to set the kube-conformance-image-version
the `Set` method ends up stripping the `v` prefix which it
requires the user to provide.

This causes a problem because the rest of the codebase assumes
the version will start with that prefix due to the fact that that
is how images are tagged (for Sonobuoy, Heptio kube-conformance,
and for upstream conformance images).

This change just adds the prefix back after validating the version
and updates the tests to actually check for the version that
gets persisted.

Fixes #716

Signed-off-by: John Schnake <jschnake@vmware.com>