Properly report error when a plugin fails to run
Currently the logic of the aggregator short circuits when a
plugin fails to `Run` without error. It exits abruptly and does
not launch other plugins or report the failure on the plugin
status.

This commit changes the following behavior:
 - ensures we can create all the auth certs needed before launching
any plugins, ensuring we don't fail to generate some certs after
other plugins are already running.
 - when a plugin fails to Run without error, we send a failing
result so that the aggregator properly reports the status of
the plugin as an error and the results tarball shows the error
as expected.

Fixes #559

Signed-off-by: John Schnake <jschnake@vmware.com>