Add ability to specify custom annotations
In some cases it may be necessary to add custom annotations
to the sonobuoy pods (aggregator and plugins). Most notably,
this helps facilitate kiam permissions.

This change adds the fields to the necessary types so that it
can be passed around appropriatly and fill in the different
templates where it should be used.

Fixes #567

Signed-off-by: John Schnake <jschnake@vmware.com>