Reprocess results if there was a processing error
In the case where a connection gets reset or other
processing errors occur, the client is already
coded to automatically retry submission of the
results.

However, the server was marking the results as 'seen'
and would not reprocess them.

This change allows reprocessing results only if there
were errors processing before.

Fixes #728

Signed-off-by: John Schnake <jschnake@vmware.com>