Add tests for job.Monitor logic
Slight refactor to improve testability of this function. It is
the first tests we have using the client-go fake client so it
sets a precedent of what/how to use it.

Helps prepare for future changes which also use the fake client
by adding the dependency on this, smaller, PR.

Signed-off-by: John Schnake <jschnake@vmware.com>