Add checks for if a container has terminated (#734)
In some cases, a plugin's main container may terminate without
reporting results. If this is the case, the sonobuoy worker
hangs forever (until the aggregator times out) because it thinks
the plugin is still running.

This change modifies the logic of the aggregator/plugin monitoring
logic in order to not only check for restarts/failure to schedule,
but also for if containers terminate without submitting results.

To ensure that we do not race with the normal plugin results
submission and termination, we did the following:
 - do not consider this a failure condition until some non-trivial
time has passed (1m set for now).
 - Stop monitoring once we see that we have results for the plugin
in question.

To facilitate those abilities, we had to make the following changes:
 - The plugin Monitor() functions now require a context which the
method uses to terminate the monitoring loop. It used to only exit
when the plugin was cleaned up.
 - The aggregator flow was modified slightly so that each plugin
would spawn its own monitor/ingestResults which would have a
dedicated channel and context.
 - Introduced a pkg/time and pkg/time/timetest package to help
control the pace of polling in those tests.

In addition, a bug was found in the recent tests that were added
that meant we were not really sleeping between polling operations
like we expected to. This has been resolved.

Fixes #667

Signed-off-by: John Schnake <jschnake@vmware.com>