Preserve single file result names and keep job/ds depth the same
If there is a job returning results it currently goes into the
tarball at plugins/name/results/<files> but this is different
from daemonset results which have an extra directory stating
the node name (plugins/name/results/node/<files>).

To prepare for consistent handling of results, it would be preferred
that these have the same layouts so we added a new level for jobs:
"global" so a job plugin results go to plugins/name/results/global.

In addition, if a single file was uploaded as results, it was handled
in a unique way. Instead of being put into the directories listed above,
the same path would refer, instead, to a file. The contents of which
were the result.

It would be much preferred to preserve the folder structure and the
filename being uploaded. This way automated processing is easier and
users can rationalize about what files they are uploading (when they
would reasonably expect them, by name, to appear in the tarball).

Fixes #741

Signed-off-by: John Schnake <jschnake@vmware.com>