Rename master to aggregator
Slightly more intutive and useful name. Kept the old name
as an alias which we can decide to remove later if we want.

Did not update a few of the config values that are named with
Master in order to avoid potentially breaking changes.

Also took this as an opportunity to slightly tweak the logic
of how the master command is run (utilizing RunE to return errors
via the cobra command flow). This is how we'd like to see the
commands structured in the future since it gets away from a global
value being used for the command.

Note, this will be backwards compatible but if someone with the
newer Sonobuoy CLI tries to target an older Sonobuoy image, then
we will fail since we provide the newer command. This is not
terribly blocking since the user just has to use an older version
of the Sonobuoy CLI.

Fixes #238

Signed-off-by: John Schnake <jschnake@vmware.com>