Add versioned docs to Sonobuoy site
Up until now we were serving the site with just a static set of
docs and it was confusing because it had duplicate data (in the
form of 'latest' version pages which were the same as 'master')
but the site also listed the verison as v0.14.2.

This change does the following:
 - adds versioning to the page
 - adds support for versions back to v0.13.0
 - fixes up links (for the main docs) to ensure they work in
this structure. Did not update the enhancement docs which are
internal and not really used/linked to.
 - fixes a bug which caused 'master' to appear in the version list
twice
 - added a warning when viewing the 'master' branch docs that they
are unstable
 - fixed a bug which caused the versioned pages to not properly
be structured due to an accidental `div` tag in the template

Fixes #738

Signed-off-by: John Schnake <jschnake@vmware.com>