Add native Makefile target
This adds a new target which can be used in the Sonobuoy Formula for
Homebrew. Homebrew formulae typically build from source. We are
currently using Docker when building and this is not available when
building for Homebrew. To ensure that we use the same build parameters
for our own builds and builds on Homebrew, we are providing a new make
target (`native`) which which runs `go build` on the host machine. This
build command is shared across the `native` target and our other build
targets.

Fixes #715

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>