Tweak the timing of considering pod as failing to be more permissive
If we are too aggressive with the timeframe, the main plugin container
may terminate and, while the sidecar is reporting the results and
the aggregator is reading them, the aggregator may instead too quickly
decide to consider the plugin a failure (because it hasn't seen the
final results yet).

This change just tries to extend some timing windows to give the
aggregator more time to see these results.

Lastly, it also adds logging into the worker when using retries. This
helps us better track what is happening on the worker's side.

Mitigates #759

Signed-off-by: John Schnake <jschnake@vmware.com>