Add basic validation of configs (#763)
This adds basic checking and validation of configs passed to some of the
interface function implementations in the Sonobuoy client. It is primarily
being added to prevent panics caused by passing nil configs to these
functions. It could be expanded in future to provide more comprehensive
validation of configuration objects.

Fixes #394

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>