Allow use of pod names other than "sonobuoy"
Currently sonobuoy always expects it's pod name to be `sonobuoy`.
So if you run sonobuoy with a CronJob, the controller will create a pod name
something like `sonobuoy-1561584180-zrwjf` and sonobuoy will fail to update
it's own status.

This change allows sonobuoy to fetch it's pod name from an environment
variable which enables using the downward API so that sonobuoy can use
whatever pod name it may be given.

If the environment variable is _not_ set, it will fall back to the
default hard-coded pod name it currently uses.

Signed-off-by: Richard Lander <landerr@vmware.com>