Look up sonobuoy master pod name by label
Currently sonobuoy always expects it's pod name to be sonobuoy.
So if you run sonobuoy with a CronJob, the controller will create a pod name
something like `sonobuoy-1561584180-zrwjf` and sonobuoy will fail to update
it's own status.  Also the sonobuoy CLI will not find the pod if you
check status or try to retrieve results.

This change removes the statically-set sonobuoy master pod name and
looks up the pod name based on the label given to that pod:
'run=sonobuoy-master`.

This assumes that this label is always set and will fail if the label is
not present.  If there are multiple pods with the label it will log a
warning and pick the first pod in the list returned.

Signed-off-by: Richard Lander <landerr@vmware.com>