Generate Plugins ConfigMap if no plugins specified (#799)
If the Sonobuoy config specified that no plugins should be run (by
overriding `Plugins` with an empty array), the ConfigMap which holds the
YAML for the plugins would not be created. The Sonobuoy aggregator would
then attempt to mount a non-existent ConfigMap and the container would
fail to start.

This change ensures that the ConfigMap is always created. In the case
where there are no plugins, no data is added.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>