Ensure only failed tests are run with --rerun-failed
This reintroduces the `e2e --rerun-failed` behaviour which had been
accidentally removed as part of a refactor. As part of reintroducing this
behaviour, it was noticed that the generated rergex for running specific
tests was invalid and would cause a panic when running ginkgo. In the
generated regex, we were double escaping `[]` characters. This doesn't
seem to be necessary as we do not double escape these characters for
other regexes (e.g. the `defaultSkipList`).

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>