Create a UUID for loaded configs if necessary (#807)
When loading a config without a UUID, we don't generate one and it
remains empty. It is necessary for the config to have a UUID as it is
used to create a uniquely named directory for storing results. Without
this, we don't correctly create a child directory to store the results.
We create a tarball within a directory that we also try to add to that
tarball, effectively adding the created tarball to itself resulting in a
write error.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>