Remove plugin.Definition
The type plugin.Definition completely duplicates fields that exist
inside of manifest.Manifest and adds nothing new. Though I prefer
the name plugin.Definition, the users currently use the manifest.Manifest
object locally and expect the nested sonobuoy-config object so we
needed to keep that in order to not break existing plugins. In addition,
the plugin.Definition type was missing the Driver field, so the server
was actually losing that data when using that type.

This change should have no impact other than clarifying the landscape
of types on the backend.

Helps facilitate #306

Signed-off-by: John Schnake <jschnake@vmware.com>