Write plugin definitions in their respective folders
The plugin info used to be written in the config but that was
removed and also didn't have any of the information like sessionID,
fields that could be altered by the config, etc.

This change dumps the loaded configuration data into the folder
at which users will most likely look for it.

In order to avoid an import cycle, it also removes a type that was
not used anywhere.

Separated this small bit of logic from my larger PR on making
plugin results more clearly reported.

Related to #306

Signed-off-by: John Schnake <jschnake@vmware.com>