Save results in a conanical format via postprocessing (#813)
Currently, when users want to see if a plugin succeeded or failed
they have to download the tarball, open the tarball, locate the
file with the results, then view them.

For the e2e tests we did introduce an `e2e` command which can parse
the results and list passed/failed/skipped tests but all the paths
are hardcoded and this isn't terribly extensible/reusable.

This change introduces a new post-processing step on the server
which will look at the results of each plugin (not just e2e) and
save the results in a standard location for each plugin. This way,
the Sonobuoy CLI (or any consumer) can automatically know where to
go to get results.

This change also introduces a `sonobuoy results` command which will
look for and parse that file to show reports to the user about
passed/failed/skipped tests of any plugin (although just junit results
are supported currently).

For non-junit results (e.g. logs) we just save a reference to each file
so that the CLI can navigate the tarball and print the results.

Fixes #306

Signed-off-by: John Schnake <jschnake@vmware.com>