Add results info to the status annotation and status cmd output
Recently we added some post-processing which would allow users
to analyze arbitrary plugins for how many tests passed/failed/etc.

This change adds some of that information into the status annotation
which we put onto the aggregator pod. This will allow people to check
results before they choose to download the tarball.

In addition, we added a bit mnore metadata into that object describing
the tarball size/name/sha which may be useful for users wanting to
confirm the data that received or save the metadata alongside the
tarball.

Fixes #819

Signed-off-by: John Schnake <jschnake@vmware.com>