Sync master/root README.md and provide a script to do this
The README for the master version of the docs and in the root
of the repo (used by Github) should always be in sync but
require minor changes due to different relative locations of
resources and links.

Also added this script as part of our CI to ensure that the two
files don't drift.

Fixes #772

Signed-off-by: John Schnake <jschnake@vmware.com>