Remove `plugin` command
This change removes the hidden `plugin` command. This command was
intended for debugging the generation of plugin Job and DaemonSet
manifests and is not in active use. There is currently other work in
progress to remove the use of templates when creating plugin resources
which means that a command to render the template is not required. If we
decided to have a debug command there would be further work required to
ensure that the same flow is used as is used in the plugin aggreator.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>