Add ability to override pod spec options
This change introduces the ability to override PodSpec options for
plugins. A new flag `--show-default-podspec` has been added to the
`gen` and `gen plugin` commands. When used, Sonobuoy will include the
default PodSpec options for the given plugin driver in the resulting
yaml. This allows users to change these values or they can provide their
own. If one is provided in the plugin definition, Sonobuoy will use that
as is, otherwise it will use the default.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>