Ensure a plugin isn't reported as passed if no results were read
In the case where the e2e plugin reported results (a tarball) but
it did not contain an xml file of junit results we want to be sure
to NOT count this as a success.

It may be reasonble to consider this a failure but I think it is
more generic/safe to just report it as unknown and users can
handle that case differently.

Also, I found another bug which was causing daemonsets to not
properly be bubbling up their results.

Fixes #840

Signed-off-by: John Schnake <jschnake@vmware.com>