Modify some old docs to avoid liquid syntax warnings
Markdown docs that tried to show go template style code like
`{{ .Foo }}` would cause liquid syntax warnings in netlify.

We always just assumed they were warnings but it actually did fail
to parse those sections so those pages are rendered incorrectly.
That is not too serious since in general we dont link to those
but it causes it to be hard to see _real_ problems in the netlify
logs and also, if you go to the page manually, shows the wrong
data.

This resolves that issue by wrapping the sections in question with
tags to make it treat those sections as 'raw' data.

In addition, there was a malformed call to `size` which I corrected.

Fixes #775

Signed-off-by: John Schnake <jschnake@vmware.com>