Use MarshalIndent on json goldenfiles
For large JSON objects stored as goldenfiles, identifying
changes is difficult when it is one large line due to the
fact that a single character change shows the entire object
as the 1 line diff.

This change makes those tests use MarshalIndent so the lines
are split up and sets the indent to 0 so that if we change the
nesting, the changes don't cascade to the rest of the file
simply due to indentation changes (which dont change the
object once unmarshalled).

Signed-off-by: John Schnake <jschnake@vmware.com>