Remove plugin.ResultType due to confusion/lack of need
There was redundancy between three different fields of the plugin:
name, resultType, and resultFormat. The first two existed prior
to the third and were the same in almost every case, leading
to confusion about why there were two different values at all.

The problem with this is that the aggregator server would be primarily
concerned with resultType and use it through its code when referring
to plugins. As a result, the plugin name, despite having the more
intuitive title, was actually almost never used.

This all led to a few confusing situations:
 - If a plugin name != resultType it was unclear on which query/screen
they would see which value
 - Uniqueness was only enforced on resultType and not plugin name, leading
to the confusing requirement that they provide a name but not using it often
and it not being required to be unique
 - The plugin name/type were almost always 100% the same even though their
names don't imply that, whereas plugin type/format were never meant to be
related despite sounding like similar ideas.

By removing the plugin result type (and relying on plugin name for the
unique identifier) we are left with two fields (name and format) which
seem much more clear and unique.

Internally, the aggregator server still uses the "ResultType" naming
convention in its own types since those types did not have the 'Name'
field to conflict with. In addition, that is not really a user facing
field whereas plugin definitions (where type/name both exist) are.

Fixes #830

Signed-off-by: John Schnake <jschnake@vmware.com>