Fix broken links and reference to conformance image
Conformance image is now maintained upstream and not in the
heptio/kube-conformance repo; fixed a reference to this.

In addition, some links to examples were in a format that would
link the docs between github *.md files but will not work on
the docs site. Updated those since the docs site is the primary
way these would be consumed.

Fixes #863

Signed-off-by: John Schnake <jschnake@vmware.com>