Adjust MASTER_URL to be just the hostname
In preparation of supporting status updates, I want to clarify
the meaning of the MASTER_URL env var on the worker.

Up until now the only URL of importance was the results URL so
that is what the MASTER_URL was set to. Soon, we will have results
and status updates, so it makes sense to have the MASTER_URL simply
be the hostname and add the path on the fly depending on which
endpoint we are trying to hit.

ref #735

Signed-off-by: John Schnake <jschnake@vmware.com>